# server/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Hub API
  description: APIs for dashboard
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
paths:
  /transmit-data:
    post:
      tags:
        - Transmission
      summary: transmit data to the server
      operationId: transmitData
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransmitDataRequest"
      responses:
        "201":
          description: Successful response
  /export-data:
    get:
      tags:
        - Dashboard
      summary: Export data within a date range
      operationId: exportData
      parameters:
        - $ref: "#/components/parameters/dateFrom"
        - $ref: "#/components/parameters/dateTo"
      responses:
        "200":
          description: CSV file containing the requested metrics
          content:
            text/csv:
              schema:
                type: string
                format: binary
...
components:
  schemas:
    EisMeasurement:
      type: object
      properties:
        containerId:
          type: string
        clusterId:
          type: string
        packId:
          type: string
        cellId:
          type: string
        realImpedance:
          type: number
          format: double
        frequency:
          type: number
          format: double
        imaginaryImpedance:
          type: number
          format: double
        voltage:
          type: number
          format: double
        creationTime:
          type: string
          format: date-time
        temperature:
          type: number
          format: double
    TransmitDataRequest:
      type: object
      properties:
        eisMeasurements:
          type: array
          items:
            $ref: "#/components/schemas/EisMeasurement"
        cellStatistics:
          type: array
          items:
            $ref: "#/components/schemas/CellStatistics"
        packStatistics:
          type: array
          items:
            $ref: "#/components/schemas/PackStatistics"
...
    CharacteristicFrequencies:
      type: object
      properties:
        impedance10Hz:
          type: number
          format: double
        impedance10000Hz:
          type: number
          format: double
...
    PackStatistics:
      type: object
      properties:
        containerId:
          type: string
        clusterId:
          type: string
        packId:
          type: string
        realTimeId:
          type: string
        maxAbsImpedance:
          type: number
          format: double
        minAbsImpedance:
          type: number
          format: double
        maxCoefficientOfVariation:
          type: number
          format: double
        minCoefficientOfVariation:
          type: number
          format: double
        maxImpedanceStdDev:
          type: number
          format: double
        minImpedanceStdDev:
          type: number
          format: double
        temperature:
          type: number
          format: double
        absImpedanceMean:
          type: number
          format: double
        impedanceStdDev:
          type: number
          format: double
        coefficientOfVariation:
          type: number
          format: double
        dispersion:
          type: number
          format: double
        degradationLevel:
          $ref: "#/components/schemas/DegradationLevel"
        suggestion:
          type: string
        alertText:
          type: string
        characteristicFrequencies:
          $ref: "#/components/schemas/CharacteristicFrequencies"
        creationTime:
          type: string
          format: date-time
        cellColors:
          type: object
          description: JSON representation of cell color information
...
    EquivalentCircuitDiagram:
      type: object
      properties:
        r0:
          type: number
          format: double
        rct:
          type: number
          format: double
        cdl:
          type: number
          format: double
        zw:
          type: number
          format: double
    Range:
      type: object
      properties:
        max:
          type: number
          format: double
        min:
          type: number
          format: double
    NyquistPlot:
      type: object
      properties:
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/NyquistPlotPoint"
        heatMap:
          $ref: "#/components/schemas/Matrix"
        xAxisRange:
          $ref: "#/components/schemas/Range"
        yAxisRange:
          $ref: "#/components/schemas/Range"
    NyquistPlotPoint:
      type: object
      properties:
        realImpedance:
          type: number
          format: double
        imaginaryImpedance:
          type: number
          format: double
    BodePlot:
      type: array
      items:
        $ref: "#/components/schemas/BodePlotPoint"
    BodePlotPoint:
      type: object
      properties:
        frequency:
          type: number
          format: double
        phase:
          type: number
          format: double
        amplitude:
          type: number
          format: double
    DrtPlot:
      type: array
      items:
        $ref: "#/components/schemas/DrtPlotPoint"
    DrtPlotPoint:
      type: object
      properties:
        relaxT:
          type: number
          format: double
        yT:
          type: number
          format: double
    Matrix:
      type: array
      items:
        type: array
        items:
          type: number
          format: double
    CellStatistics:
      type: object
      properties:
        containerId:
          type: string
        clusterId:
          type: string
        packId:
          type: string
        groupId:
          type: string
        cellId:
          type: string
        realTimeId:
          type: string
        absImpedanceMean:
          type: number
          format: double
        impedanceStdDev:
          type: number
          format: double
        coefficientOfVariation:
          type: number
          format: double
        phase:
          type: number
          format: double
        impedanceMeanToMaxRatio:
          type: number
          format: double
        impedanceMeanToMinRatio:
          type: number
          format: double
        impedanceMeanToAvgRatio:
          type: number
          format: double
        impedanceStdDevToMaxRatio:
          type: number
          format: double
        impedanceStdDevToMinRatio:
          type: number
          format: double
        impedanceStdDevToAvgRatio:
          type: number
          format: double
        bodePlot:
          $ref: "#/components/schemas/BodePlot"
        nyquistPlot:
          $ref: "#/components/schemas/NyquistPlot"
        drtPlot:
          $ref: "#/components/schemas/DrtPlot"
        realPartCorrelation:
          $ref: "#/components/schemas/Matrix"
        imagPartCorrelation:
          $ref: "#/components/schemas/Matrix"
        equivalentCircuitDiagram:
          $ref: "#/components/schemas/EquivalentCircuitDiagram"
        creationTime:
          type: string
          format: date-time
        realPart10Hz:
          type: number
          format: double
        imaginaryPart10Hz:
          type: number
          format: double
        maxImpedanceStdDev:
          type: number
          format: double
        minImpedanceStdDev:
          type: number
          format: double
        maxAbsImpedance:
          type: number
          format: double
        minAbsImpedance:
          type: number
          format: double
        maxCoefficientOfVariation:
          type: number
          format: double
        minCoefficientOfVariation:
          type: number
          format: double
...
    DegradationLevel:
      type: string
      enum: ["Normal", "Mild", "Severe"]
  parameters:
...
    dateFrom:
      in: query
      schema:
        type: string
        format: date-time
      required: false
      name: dateFrom
    dateTo:
      in: query
      schema:
        type: string
        format: date-time
      required: false
      name: dateTo